+++ client/src/100_element/circle/sphere/manifold/[sphere_manifold].tsx
