+++ client/src/100_element/circle/sphere/manifold/star.tsx
