+++ client/src/100_element/manifold/[manifold].tsx
