+++ client/src/300_entity/atmosphere/atmosphere.tsx
